[jit][edge] Export maybe-used interface methods from modules.
Differential Revision: [D31262117](https://our.internmc.facebook.com/intern/diff/D31262117/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D31262117/)!

[ghstack-poisoned]
